Add passport-jwt strategy

+++ src/tvhelper/db/models/FavoritesModel.js
@@ -4,6 +4,11 @@ import { Schema } from 'mongoose';
 
 import mongoose from '../../../common/db/MongoDB';
 
+type Favorite = {|
+  +userId: ?string,
+  +serieId: string,
+|};
+
 const FavoritesSchema = new Schema({
   userId: {
     type: Schema.Types.ObjectId,
@@ -18,15 +23,28 @@ FavoritesSchema.index({ userId: 1, serieId: -1 }, { unique: true });
 
 const Favorites = mongoose.model('favorites', FavoritesSchema);
 
-export const createFavorite = (favorite: Object) =>
-  Favorites.create(new Favorites(favorite));
+export const createFavorite = (favorite: Favorite) => {
+  if (favorite.userId == null) {
+    throw Error('You must be logged in to add favorite.');
+  }
+  return Favorites.create(new Favorites(favorite));
+};
 
-export const deleteFavorite = ({ userId, serieId }: Object) =>
-  Favorites.deleteOne({
+export const deleteFavorite = ({ userId, serieId }: Favorite) => {
+  if (userId == null) {
+    throw Error('You must be logged in to delete favorite.');
+  }
+  return Favorites.deleteOne({
     userId,
     serieId,
   });
+};
 
-export const findFavorites = (userId: string) => Favorites.find({ userId });
+export const findFavorites = (userId?: string) => {
+  if (userId == null) {
+    throw Error('You must be logged in to get favorites');
+  }
+  return Favorites.find({ userId });
+};
 
 export default Favorites;
